rh-assistant/backend/requirements.txt

fastapi
uvicorn
python-dotenv
openai
chromadb
sentence-transformers
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
SQLAlchemy
alembic
redis
aioredis
pydantic-settings
loguru
python-multipart

